Fix error calling SB-POSIX:STAT on a file stream fd.
SBCL has an error on windows where calling
(SB-POSIX:STAT (SB-SYS:FD-STREAM-FD <stream>)) yields an error
"The storage control block address is invalid. (9)". This is likely a bug
in SBCL, but it's easy to avoid here, by retrieving a pathname with
(TRUENAME <stream>) and passing that to STAT instead. This change allows
archive to successfully create tar files on my machine.